+++ test/vips_test.rb
@@ -53,6 +53,9 @@ describe "ImageProcessing::Vips" do
 
     result = ImageProcessing::Vips.loader(autorot: false).call(fixture_image("rotated.jpg"))
     assert_dimensions [800, 600], result
+
+    result = ImageProcessing::Vips.loader(autorotate: false).call(fixture_image("rotated.jpg"))
+    assert_dimensions [800, 600], result
   end
 
   it "accepts Vips::Image as source" do